DRF-TokenBasedAuth/api/serializers.py
from rest_framework import serializers
from accounts.models import MyUser

class MyUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = MyUser
        fields = ['email', 'first_name', 'last_name', 'username', 'password']